Avoid pausing for logging only if poll_status() is null (h48 solver).
This is a small amendment to c6a77f30f64be73a5e55e06336975f2ecfbb2324,
which changed the way we log solutions while the h48 solver is running.
With the method recently introduced, the main thread checks for solutions
to log every 0.5 seconds, resulting in a possible slowdown of at most
0.5s per solve. The solutions are also logged when all worker threads are
completed. With this new method, when the poll_status() callback function
is NULL, which likely means nissy is not run interactively, we rely
only on the final log on completion of the worker threads. This means
less frequent logging, but at no performance cost.