Minor improvement
Add a trivial check to H48 solver. The benchmarks show, surprisingly,
a more noticeable improvement for larger solvers (h10, h11). I was expecting
it to be more noticeable for smaller solvers.

It is likely that this change is irrelevant to performance and all changes
in the benchmark results are just random chance.